add an extra check for file existence when creating a file

+++ create_bash_script_and_make_executable.sh
@@ -5,9 +5,14 @@ fileName=$1
 if [[ -n $fileName ]];
 then
   touch $fileName
-  chmod +x $fileName
-  printf "#!/usr/bin/env bash\n\n\n" > $fileName
-  vim $fileName
+  if [[ -f $fileName ]];
+  then
+    chmod +x $fileName
+    printf "#!/usr/bin/env bash\n\n\n" > $fileName
+    vim $fileName
+  else
+    echo "Failed to create file"
+  fi
 else
   echo "Please provide a file name"
 fi
